fixup! feat(cascade): serve records (not ops) on the data-serving wire
